Better examine script

+++ examine.sh
@@ -3,11 +3,18 @@
 # examine.sh depends on go, sakemake and bat
 #
 go build
-fn=$(go test | grep "FAILED:" | cut -d" " -f4)
+fn=$(go test | grep "Compiling and running" | cut -d"]" -f2- | cut -d" " -f5 | tail -1)
+if [ ! -e "$fn" ]; then
+  echo "No issues."
+  exit 0
+fi
 echo 'Go program:'
-bat "$fn"
+bat "$fn" || exit 1
 echo 'C++17 program:'
 mkdir -p test
 ./go2cpp "$fn" -o | tee test/main.cpp | bat -l cpp
+echo "Go output:"
 go run "$fn"
-grep -q main test/main.cpp && sm -C test/ run || echo 'No C++ program to run.'
+grep -q main test/main.cpp && sm -C test >/dev/null || exit 1
+echo "C++ output:"
+sm -C test run
